Add argparse cli
This replaces a bunch of features which were previously "please edit the
python file to enable" with command line flags. Where relevant command
line arguments will override config file settings. Notable new arguments
include:
--action which sets normal or test modes
--mock which selectively disables external hardware/network dependencies
See --help for more.

I tried to leave the rest of the structure and behavior as is except in
the following areas:
- Moved config file reading to a function and moved associated global
  variables to locals/the new Config class. It was surprising that it
  tried to set up a config or read it when just importing main.py. Now
  it will only do that when calling main(). The previous scattering of
  global variables are now collected in Config and passed as needed to
  consumers.
- Removed the default arguments from Main(). Default arguments in python
  are a little bit foot-guny in that they get created once at import
  time instead of new instances for each invocation.