Convert config to TOML
Toml is a well specified and widely supported configuration language
similar to ini. Among other things it supports nested structures
allowing the local TLE cache to be merged in. This also no longer
automatically generates a config template, instead prompting the user to
make one with the `--template` flag.

Note that when upgrading in place pass_commander installs,
`pip install -r requirements.txt` will be needed to be rerun,
pass_commander.ini should be renamed to pass_commander.toml, and in the
config double quotes must be added to all values interpreted as a
string.