Preserve lat/lon as strings
Setting lat/lon in Observer is done by an implicit conversion to
ephem.Angle. Angle is internally always a radian float and normally the
library only lets you construct Angles directly via ephem.degrees() or
ephem.hours(). Setting Observer.lat/lon bypasses the constructors and
with a string will parse it correctly as a degrees string, but setting
it with a degrees float will be instead interpreted as radian float.

Here we revert to the original behavior of strings to preserve the
previous functionality instead of converting to radians.