Change lat/lon config fields to numeric type
Is it an act of hubris? Rebellion? This mostly reverts f48d12c7100 which
changed lat/lon back to strings because a bugs. This time it will be
different surely.

Since toml has types it feels kinda awkward to give numbers as strings.
As long as it's ensured that the degrees are converted to radians before
being converted to ephem.Angles it should be fine, that was the downfall
last time. This does mean though that we won't accept 'h:m:s' notation,
but we never really thought we could accept it anyway.