Move config template text to template()
Instead of dataclass default values. It makes the types for the
dataclass make more sense, having a Union[Numeric, str] for lat/lon/alt
for example was confusing because the str was only for the template,
which is discarded after creation.

This also adds some additional type annotations which makes mypy happy.