Add packet getter, shim packet buffer
For reasons that I increasingly don't understand we want a Front End
Processor, or FEP, to sit between Pass Commander and a source of packets
and it must be a separate process.

Pass Commander would like to control exactly when packets get sent and
to do that it through some undecided mechanism fetches them one by one
from the FEP when it is ready. Unfortunately none of the rest of our
infrastructure is built this way so this patch introduces a shim which
will buffer UDP EDL packets from the rest of our tools and provide and
XMLRPC interface for Pass Commander to fetch them one by one from. It
also updates Pass commander to fetch them one by one via XMLRPC. XMLRPC
was chosen because it is already being used elsewhere in the project but
it's easy to change protocol if so desired.

To easily accommodate this I changed PTT behavior to always be on while
the pass is above the minimum elevation. As far as I understand this
should be fine assuming we are using full-duplex lband/uhf. We don't
plan on using half-duplex uhf for the forseeable future.

This also removes manual EDL packet configuration from Pass Commander
itself, instead moving it into the shim. This is keeping in line with
having as little difference between live and testing code of Pass
Commander as possible.