Add mock_station.py, improve mock_flowgraph.py
I'd like to run a full real pass with no control flow changes in pass
commander, so that means using external mocked stubs of the external
dependencies. This starts the process for stationd, which has a complex
protocol that I did the minimum effort to sim, adds a few missing
commands from the mock flowgraph, and adds an edl listener to the
flowgraph.

The rotator and pass selection still needs to be done but this is a
start.