Replace pydbus with jeepney
While pydbus is nice, it relies on system libraries and can't be
installed from pypi. That makes this project harder to package and
develop, with things like venvs. The library jeepney seems to be
the premier python only dbus binding and, while a little more
complicated in use, seems to be mostly a drop in replacement.

The real solution to this problem is not to use dbus and just call
adjtimex() directly but I can't find python bindings for it.