Setup python packaging
This uses the fairly standard distutils/build process for generating
wheels. The one major hangup is that Hamlib isn't on pypi so it can't be
captured in the dependencies. I'd ideally like to produce a debian
package for that but my brain is too smooth to comprehend the full
process.