Add UHF LNA control
This adds stationd control for the UHF LNA, which is a weird RF relay
thing that needs to be switched in just the right way for it to actually
work.

Also split the start() function in main up partly into class Main and
partly into main(). All the setup is now in class Main and the remaining
bits are picking what action, which could conceivably be part of arg
handling in main().