Port from pyephem to skyfield
A bit of an understatement really, it is more like a broad rewrite.
Skyfield is by the same author as pyephem but is a re-imagination to
correct all the flaws of the original, and also pyephem is no longer
maintained.

The main two new features of skyfield that are used here are the units
(Angle, Velocity, Time) and vector operations. Units are nice because
they attempt to provide some type level enforcement and vector
operations on those units means we can calculate the azimuth, elevation,
and range velocities of a satellite pass in advance and then step
through them at the appropriate time instead of waiting for the correct
point in time to calculate them on the fly. With that we could
eventually validate the entire rotator navigation beforehand against
given physical limits or run the whole pass in compressed time for
testing purposes.

Because we now have a bunch of time/position entries then the main pass
runner can be written in an event driven style where a timer is set for
each interesting time and an associated callback then gets invoked, for
example to set the rotator position or doppler shift. One day this could
probably be rewritten again to be fully python async.

While doing this I wrote a bunch of mock testing infrastructure and unit
tests to support this. Mocking hamlib was hard though, and given the
packaging woes I opted to swap it out for direct management of the
rotator using the rot2prog package. With hamlib gone pass-commander no
longer has any system package dependencies and for example could be
installed from pypi with no extra steps.