Create rotcmd.py, a small cmdline to interact with the rotator
This is a small util for testing rotator commands through the
rot2sim/Rotator interface. With hamlib/rotctl going away we need a
simple way for operators to manually interact with the rotator.

To test the tester mock.rotator can now be run directly. It will set up
the mock rotator and print the path to interact with it from.