Rotator timeout fix, hardware cleanup, logging fixes
When sending commands to the rotator it turns out it doesn't like
messages sent too fast to it and will ignore the second message. This
causes an error which stops passes so this also puts the hardware
teardown in a finally block to make it more likely to be run.

Finally a number of log messages have been made more readable and
timestamps have been added.