Fix infinite recursion on PA error
In the real stationd, it errors when trying to turn the PA off with PTT
on. Pass commander's behavior was to immediately recursively retry the
action, which lead to a recursion overflow. This fixes that by turning
the error handling from retry to exception. It's a programming error to
turn the PA off in the wrong order so we should know immediately.

In order to test this mock.station needed to be enhanced with some state
tracking and the tests were cleaned up and extended to check the cases
where PTT conflicts would occur.