Decrease rotator command rate, fix ptt order
It turns out the rot2prog rotator controller can only be commanded once
every two seconds or it could experience *undefined behavior*. This will
eventually necessitate a whole rethink of how it gets operated but for
now these are a couple quick hacks in preparation for the next pass.

Also adds a ptt_off() in the SinglePass.work() cleanup that was missing
and could lead to stationd erroring about trying to turn the pa off
with ptt on.