Sleep until end of pass when thermal limit exceeded
Hey look there was even a fixme for it. We hit this today while running
a daytime pass, when the thermal limit is exceeded previously it would
start another iteration immediately, see that we were in a pass, and
then check the temps again and fail again.

The new behavior is to sleep until the end of the current pass when the
thermal limits are exceeded so we don't attempt this pass again.