Separate rotator positioning from polling
Since polling the rotator can't happen during a pass, but there's still
moments when we want to poll its position, that functionality was split
into Rotator.wait_for(). Now that I write this here the name could
probably better, it's not doing any blocking. In SinglePass the old way
of monitoring Rotator.listener was removed because it doesn't work like
that anymore.

Also the unit test for main() was removed temporarily, it turns out
pytest doesn't reset argv per test, it'll just pass the argv it was
invoked with along.