Add logging for absolute position of the satellite
Previously when calculating the rotator position in realtime we'd log
both the satellite position and the rotator position. When moving to
ahead of time calculation the satellite position got dropped but it
turns out it's nice information for the operator to see. This adds a
callback that will log pos at the appropriate times.