Use ephemeral ports for mock dependencies
It was always a little awkward specifying the mock dependency socket
addresses by hand and having some on 127.0.0.2. Additionally if they
were run on a live station with the real staitond running, which binds
to all addresses, the ports would have to be changed as well. Instead of
coming up with more config options to deconflict rare cases it occurred
to me I could ask the kernel to solve all my problems with ephemeral
ports. If the kernel runs out of those we've got bigger problems.

As an added benefit this helps test that things aren't dependant on
hardcoded port numbers.

Finally I also fixed up a bit of formatting and cleaned up/expanded some
comments.