Handle stale TLEs better
This concentrates TLE fetching and pass planning into
sleep_until_next_pass() and tries to be a bit nicer about fetching the
TLEs. They get updated once or twice a day typically and it's nice to
not fetch them too often (API limits). Previously it would fetch and
then sleep and then fetch again, even if we had only slept a couple
seconds.