Add configurable max temperature
Now that it's summer it's the perfect time to make the thermal limit
configurable. During a demonstration a couple weeks ago we ran into
the need for this where the in-box temperature was just slightly over
the limit and we the operators were willing to just override it.

This adds both a command line option and a config file option. The
default value has also been raised from 30 to 40 on recommendation from
Glenn.