Add pointing mode
This is a testing feature, the intent is someone on the ground with an
engineering unit satellite to go to a specific location and communicate
that coordinate to the station operator. This will point the antenna at
that coordinate and then run all the actions for a normal pass but
indefinitely until the operator stops operation with ctrl+c. Conflicts
with normal operation but it's not yet enforced in the command line
invocation.