Check station temperature during a pass
In talking with Glenn it turns out he expects the PAs to noticeably heat
up the enclosure so we should make sure during the entire operation that
we're within the expected thermal limits. This checks every 30 seconds
which I assume is fine but maybe we can adjust this in practice.